+++ src/main/java/core/mvc/tobe/handler/mapping/NoExistsHandlerException.java
@@ -1,4 +1,4 @@
-package core.mvc.tobe;
+package core.mvc.tobe.handler.mapping;
 
 import java.util.NoSuchElementException;
