+++ src/test/java/core/mvc/tobe/handler/mapping/HandlerMappingsTest.java
@@ -1,4 +1,4 @@
-package core.mvc.tobe;
+package core.mvc.tobe.handler.mapping;
 
 import core.mvc.asis.Controller;
 import org.junit.jupiter.api.DisplayName;
